+++ src/main/java/lotto/domain/BonusRequirement/BonusRequirement.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.BonusRequirement;
 
 import java.util.function.Function;
