+++ src/test/java/lotto/domain/BonusRequirement/BonusRequirementExpressionsTest.java
@@ -1,11 +1,9 @@
-package lotto.domain;
+package lotto.domain.BonusRequirement;
 
+import lotto.domain.BonusRequirement.BonusRequirementExpressions;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class BonusRequirementExpressionsTest {
