+++ src/test/java/lotto/domain/BonusRequirement/BonusRequirementTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.BonusRequirement;
 
+import lotto.domain.BonusRequirement.BonusRequirement;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
